+++ CHANGELOG.md
@@ -89,6 +89,9 @@
   [Cihat Gündüz](https://github.com/Dschee)
   [#1587](https://github.com/realm/SwiftLint/issues/1587)
 
+* Add opt-in rules for Firebase SDK.  
+  [Ibrahim Ulukaya (Google Inc.)](https://github.com/iulukaya)
+
 ##### Bug Fixes
 
 * Fix false positive in `empty_enum_arguments` rule when calling methods.